requirements/dev.txt
-r base.txt

# pytest requirements
attrs==21.4.0
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.6
pytest==6.2.5
toml==0.10.2

# pylint requirements
astroid==2.9.1
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
platformdirs==2.4.1
pylint==2.12.2
typing-extensions==4.0.1
wrapt==1.13.3

# fabric requirements
bcrypt==3.2.0
cffi==1.15.0
cryptography==36.0.1
fabric==2.6.0
invoke==1.6.0
paramiko==2.9.1
pathlib2==2.3.6
pynacl==1.4.0

# Debugging requirements
django-debug-toolbar==3.2.4